Suppress conversion warnings reported by MSVC.

+++ asio/include/asio/ssl/impl/context.ipp
@@ -427,7 +427,8 @@ asio::error_code context::use_certificate(
 
   if (format == context_base::asn1)
   {
-    if (::SSL_CTX_use_certificate_ASN1(handle_, buffer_size(certificate),
+    if (::SSL_CTX_use_certificate_ASN1(handle_,
+          static_cast<int>(buffer_size(certificate)),
           buffer_cast<const unsigned char*>(certificate)) == 1)
     {
       ec = asio::error_code();
@@ -928,7 +929,7 @@ BIO* context::make_buffer_bio(const const_buffer& b)
 {
   return ::BIO_new_mem_buf(
       const_cast<void*>(buffer_cast<const void*>(b)),
-      buffer_size(b));
+      static_cast<int>(buffer_size(b)));
 }
 
 #endif // !defined(ASIO_ENABLE_OLD_SSL)
